Development

+++ src/controllers/users.controller.js
@@ -4,17 +4,37 @@ const Users = db.users;
 exports.findAll = (req, res) => {
   Users.find()
     .then((result) => {
-      res.send(result);
+      res.send({
+        name: result.name,
+        username: result.username,
+        email: result.email,
+        image: result.imageLink,
+        type: {
+          memberType: result.type.accountType.member,
+          adminType: result.type.isAdmin,
+        },
+        referal: {
+          referalCode: result.referal.referalCode,
+          referalCount: result.referal.referalCount,
+          referalAccount: result.referal.referalAccount,
+        },
+      });
     })
     .catch((err) => {
       res.status(500).send({
-        message: err.message || "Some error while retrieving users.",
+        message: err.message || "Some error occurred while fetching the Users.",
       });
     });
 };
 
 exports.findOne = (req, res) => {
   const id = req.params.id;
+
+  if (!id) {
+    return res.status(400).send({
+      message: "Id is required",
+    });
+  }
 
   Users.findById(id)
     .then((result) => {
@@ -23,7 +43,22 @@ exports.findOne = (req, res) => {
           message: "User not found",
         });
       }
-      res.send(result);
+
+      res.send({
+        name: result.name,
+        username: result.username,
+        email: result.email,
+        image: result.imageLink,
+        type: {
+          memberType: result.type.accountType.member,
+          adminType: result.type.isAdmin,
+        },
+        referal: {
+          referalCode: result.referal.referalCode,
+          referalCount: result.referal.referalCount,
+          referalAccount: result.referal.referalAccount,
+        },
+      });
     })
     .catch((err) => {
       res.status(500).send({
@@ -34,6 +69,12 @@ exports.findOne = (req, res) => {
 
 exports.delete = (req, res) => {
   const id = req.params.id;
+
+  if (!id) {
+    return res.status(400).send({
+      message: "Id is required",
+    });
+  }
 
   Users.findByIdAndRemove(id)
     .then((result) => {
@@ -54,6 +95,12 @@ exports.delete = (req, res) => {
 exports.update = (req, res) => {
   const id = req.params.id;
 
+  if (!id) {
+    return res.status(400).send({
+      message: "Id is required",
+    });
+  }
+
   Users.findByIdAndUpdate(id, req.body)
     .then((result) => {
       if (!result) {
@@ -61,11 +108,88 @@ exports.update = (req, res) => {
           message: "User not found",
         });
       }
-      res.send(result);
+      res.send({
+        message: "User updated successfully.",
+        data: {
+          name: result.name,
+          username: result.username,
+          email: result.email,
+          image: result.imageLink,
+          type: {
+            memberType: result.type.accountType.member,
+            adminType: result.type.isAdmin,
+          },
+          referal: {
+            referalCode: result.referal.referalCode,
+            referalCount: result.referal.referalCount,
+            referalAccount: result.referal.referalAccount,
+          },
+        },
+      });
     })
     .catch((err) => {
       res.status(500).send({
         message: err.message || "Some error occurred while updating the user.",
       });
     });
+};
+
+exports.changePassword = (req, res) => {
+  const id = req.params.id;
+  const oldPassword = req.body.oldPassword;
+  let newPassword = req.body.newPassword;
+
+  if (!id || !oldPassword || !newPassword) {
+    return res.status(400).send({
+      message: "Id, oldPassword and newPassword are required",
+    });
+  }
+
+  const user = Users.findById(id);
+  if (!user) {
+    return res.status(404).send({
+      message: "User not found",
+    });
+  }
+
+  bcrypt.compare(oldPassword, user.password, (err, result) => {
+    if (err) {
+      return res.status(500).send({
+        message:
+          err.message || "Some error occurred while changing the password.",
+      });
+    } else if (!result) {
+      return res.status(401).send({
+        message: "Invalid password",
+      });
+    } else {
+      bcrypt.hash(newPassword, 10, (err, hash) => {
+        if (err) {
+          return res.status(500).send({
+            message:
+              err.message || "Some error occurred while changing the password.",
+          });
+        }
+        newPassword = hash;
+        Users.findByIdAndUpdate(id, { password: newPassword })
+          .then((result) => {
+            if (!result) {
+              res.status(404).send({
+                message: "User not found",
+              });
+            }
+            res.send({
+              message: "User's password updated successfully.",
+            });
+          })
+          .catch((err) => {
+            res.status(500).send({
+              message:
+                err.message ||
+                "Some error occurred while changing the password.",
+            });
+          });
+      });
+    }
+  });
 };
